feat(): Added bottom tab navigation and dependent modules
