+++ babel.config.js
@@ -2,6 +2,7 @@ module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
   plugins: [
     'nativewind/babel',
+    'react-native-reanimated/plugin',
     [
       'module-resolver',
       {
