+++ package.json
@@ -31,6 +31,8 @@
     "react-native-paper": "^5.9.1",
     "react-native-paper-dates": "^0.18.12",
     "react-native-reanimated": "^3.3.0",
+    "react-native-safe-area-context": "^4.7.1",
+    "react-native-screens": "^3.22.1",
     "react-native-toast-message": "^2.1.6",
     "react-native-vector-icons": "^9.2.0",
     "react-redux": "^8.1.1",
@@ -44,6 +46,7 @@
     "@react-native/metro-config": "^0.72.9",
     "@tsconfig/react-native": "^3.0.0",
     "@types/react": "^18.0.24",
+    "@types/react-native-vector-icons": "^6.4.13",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.2.1",
     "babel-plugin-module-resolver": "^5.0.0",
