+++ src/screens/Performance/performance.screen.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+export default function PerformanceScreen(): JSX.Element {
+  return (
+    <View>
+      <Text>Performance Screen</Text>
+    </View>
+  );
+}
