+++ src/screens/Profile/profile.screen.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {View, Text} from 'react-native';
+
+export default function ProfileScreen(): JSX.Element {
+  return (
+    <View>
+      <Text>Profile Screen</Text>
+    </View>
+  );
+}
