+++ src/screens/main.screen.tsx
@@ -1,155 +1,126 @@
-import React, {useEffect} from 'react';
-import {ScrollView, Text} from 'react-native';
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import React from 'react';
+import {View, StyleSheet, Animated, Dimensions} from 'react-native';
+import MaterialIcon from 'react-native-vector-icons/MaterialCommunityIcons';
+import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 
-import auth from '@react-native-firebase/auth';
+const Tab = createBottomTabNavigator();
 
-import useFirebase from '@/hooks/firebase.auth.hook';
-import {useSelector} from 'react-redux';
-import {RootState} from '@/redux/store';
+import HomeScreen from './Home/home.screen';
+import ProfileScreen from './Profile/profile.screen';
+import PerformanceScreen from './Performance/performance.screen';
 
-// import SafeArea from '@components/SafeArea';
-import SafeArea from '@/components/SafeArea';
-import {MyToast} from '@/components/MyToast';
-import {Button} from 'react-native-paper';
-import {useGetAllExerciseDetails} from '@/apis/exercise.db';
-import {useExtractQuery} from '@/hooks/useExtractFirebaseData';
-import Toast from 'react-native-toast-message';
-import GlobalLoadingView from '@/components/GlobalLoadingView';
+const styles = StyleSheet.create({
+  shadow: {
+    shadowColor: '#7F5DF0',
+    shadowOffset: {
+      width: 0,
+      height: 10,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.5,
+    elevation: 5,
+  },
+});
+
+const GenIcon = (focused: boolean, name: string) => (
+  <View className="d-flex flex-col h-full absolute top-1/2">
+    <MaterialIcon
+      name={name}
+      size={30}
+      color={focused ? '#e32f45' : '#748c94'}
+    />
+  </View>
+);
+
+const getWidth = () => {
+  let width = Dimensions.get('window').width;
+
+  width = width - 40;
+
+  return width / 3;
+};
 
 export default function MainScreen(): JSX.Element {
-  const authState = useSelector((state: RootState) => state.auth);
-  const {getCurrentUser, getTokens, signIn, isSignedIn, signOut} =
-    useFirebase();
-
-  const {handleAuthStateChanged} = useFirebase();
-
-  const {data, isLoading, isError, error, refetch, fetchStatus} =
-    useGetAllExerciseDetails();
-  const {mappedData} = useExtractQuery(data);
-  const {LoadingView, setIsLoading} = GlobalLoadingView();
-
-  if (!isLoading && !isError) {
-    console.log('Extracted Data: ', mappedData);
-  }
-
-  useEffect(() => {
-    if (!isLoading && error) {
-      Toast.show({
-        type: 'error',
-        text1: 'ERROR ',
-        text2: `${error.message}`,
-      });
-    }
-    if (isLoading && fetchStatus === 'fetching') {
-      console.log('Should show spinner!');
-      setIsLoading(true);
-    } else {
-      setIsLoading(false);
-    }
-  }, [error, isLoading, fetchStatus, setIsLoading]);
-
-  const backgroundStyle = 'bg-transparent dark:bg-slate-900';
-
-  useEffect(() => {
-    const subscriber = auth().onAuthStateChanged(handleAuthStateChanged);
-    return subscriber; // unsubscribe on unmount
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
+  const tabOffsetValue = React.useRef(new Animated.Value(getWidth())).current;
   return (
     <>
-      <LoadingView />
-      <SafeArea>
-        <ScrollView
-          contentInsetAdjustmentBehavior="automatic"
-          className={backgroundStyle}>
-          <Button
-            icon="camera"
-            mode="contained"
-            onPress={() => console.log('Pressed')}>
-            Press me
-          </Button>
-          {/* {initializing && <Text>Initializing Firebase...</Text>} */}
-          {authState.isAuthenticated ? (
-            <>
-              <Button onPress={() => getTokens()}>Get Tokens</Button>
-              <Button onPress={() => isSignedIn()}>Is Signed In?</Button>
-              <Button onPress={() => getCurrentUser()}>Get Current User</Button>
-              <Button onPress={() => refetch()}>Fetch Exercise Data</Button>
-              <Button onPress={() => signOut()}>Sign Out</Button>
-            </>
-          ) : (
-            <>
-              <Button
-                icon="google"
-                mode="contained-tonal"
-                onPress={() => signIn()}>
-                Google SignIn
-              </Button>
-              <Button onPress={() => isSignedIn()}>Is Signed In?</Button>
-              <Button onPress={() => getCurrentUser()}>Get Current User</Button>
-              <Button onPress={() => refetch()}>Fetch Exercise Data</Button>
-              <Button onPress={() => getTokens()}>Get Tokens</Button>
-              <Text>Sign in Status: Not Signed In</Text>
-            </>
-          )}
-
-          <Button
-            onPress={() => {
-              setIsLoading(true);
-              Toast.hide();
-              Toast.show({
-                type: 'error',
-                text1: 'Error',
-                autoHide: true,
-                text2:
-                  'This is some something 👋 [Error: [firestore/permission-denied] The caller does not have permission to execute the specified operation.]',
-              });
-            }}>
-            Test Error Toast
-          </Button>
-
-          <Button
-            onPress={() =>
-              Toast.show({
-                type: 'info',
-                text1: 'Info',
-                text2:
-                  'This is some something 👋 [Error: [firestore/permission-denied] The caller does not have permission to execute the specified operation.]',
-              })
-            }>
-            Test Info Toast
-          </Button>
-
-          <Button
-            onPress={() =>
-              Toast.show({
-                type: 'warning',
-                text1: 'Warning',
-                text2:
-                  'This is some something 👋 [Error: [firestore/permission-denied] The caller does not have permission to execute the specified operation.]',
-              })
-            }>
-            Test Warning Toast
-          </Button>
-
-          <Button
-            onPress={() =>
-              Toast.show({
-                type: 'success',
-                text1: 'Success',
-                text2:
-                  'This is some something 👋 [Error: [firestore/permission-denied] The caller does not have permission to execute the specified operation.]',
-              })
-            }>
-            Test Success Toast
-          </Button>
-          <Text>REDUX STATE: {JSON.stringify(authState)}</Text>
-
-          {/* Global Toast */}
-          <MyToast />
-        </ScrollView>
-      </SafeArea>
+      <Tab.Navigator
+        initialRouteName="Home"
+        screenOptions={{
+          tabBarShowLabel: false,
+          tabBarStyle: {
+            position: 'absolute',
+            bottom: 40,
+            left: 20,
+            right: 20,
+            height: 70,
+            borderRadius: 15,
+            ...styles.shadow,
+          },
+        }}>
+        <Tab.Screen
+          name="Performance"
+          component={PerformanceScreen}
+          options={{
+            tabBarIcon: ({focused}) => GenIcon(focused, 'poll'),
+          }}
+          listeners={({navigation, route}) => ({
+            tabPress: e => {
+              Animated.spring(tabOffsetValue, {
+                toValue: 0,
+                useNativeDriver: true,
+              }).start();
+            },
+          })}
+        />
+        <Tab.Screen
+          name="Home"
+          component={HomeScreen}
+          options={{
+            tabBarIcon: ({focused}) => GenIcon(focused, 'home'),
+          }}
+          listeners={({navigation, route}) => ({
+            tabPress: e => {
+              Animated.spring(tabOffsetValue, {
+                toValue: getWidth(),
+                useNativeDriver: true,
+              }).start();
+            },
+          })}
+        />
+        <Tab.Screen
+          name="Profile"
+          component={ProfileScreen}
+          options={{
+            tabBarIcon: ({focused}) => GenIcon(focused, 'account'),
+          }}
+          listeners={({navigation, route}) => ({
+            tabPress: e => {
+              Animated.spring(tabOffsetValue, {
+                toValue: getWidth() * 2,
+                useNativeDriver: true,
+              }).start();
+            },
+          })}
+        />
+      </Tab.Navigator>
+      <Animated.View
+        // eslint-disable-next-line react-native/no-inline-styles
+        style={{
+          position: 'absolute',
+          bottom: 40,
+          left: 40,
+          height: 2,
+          width: getWidth() - 40,
+          backgroundColor: 'red',
+          transform: [
+            {
+              translateX: tabOffsetValue,
+            },
+          ],
+        }}
+      />
     </>
   );
 }
